+++ src/Rpg.Logic/Common/Constants/AlchemyConstants.cs
@@ -2,8 +2,8 @@ namespace Rpg.Logic.Common.Constants
 {
     public class AlchemyConstants
     {
-        public const double IngredientMultiplier = 4.0;
+        public const decimal IngredientMultiplier = 4.0M;
 
-        public const double AlchemySkillFactor = 1.5;
+        public const decimal AlchemySkillFactor = 1.5M;
     }
 }
