+++ tests/Rpg.Logic.Tests/Alchemy/PotionFactoryTests/Base.cs
@@ -52,15 +52,15 @@ namespace Rpg.Logic.Tests.Alchemy.PotionFactoryTests
             AlchemyEffectFactory
                 .Setup(x => x.Create(
                     It.IsAny<IAlchemyEffect>(),
-                    It.IsAny<double>(),
-                    It.IsAny<double>(),
-                    It.IsAny<double>()
+                    It.IsAny<decimal>(),
+                    It.IsAny<decimal>(),
+                    It.IsAny<decimal>()
                 ))
                 .Returns((
                     IAlchemyEffect effect,
-                    double cost, 
-                    double duration,
-                    double magnitude) => 
+                    decimal cost, 
+                    decimal duration,
+                    decimal magnitude) => 
                     {
                         var mock = new Mock<IAlchemyEffect>();
                         mock.Setup(x => x.Cost).Returns(cost);
